(PC-36371)[PRO] refactor: split IndividualOfferDetails serializers into another file
