+++ pro/src/pages/IndividualOffer/IndividualOfferDetails/commons/utils.ts
@@ -10,11 +10,11 @@ import { showOptionsTree } from 'commons/core/Offers/categoriesSubTypes'
 import { isOfferSynchronized } from 'commons/core/Offers/utils/typology'
 import { SelectOption } from 'commons/custom_types/form'
 import { getAccessibilityInfoFromVenue } from 'commons/utils/getAccessibilityInfoFromVenue'
-import { trimStringsInObject } from 'commons/utils/trimStringsInObject'
 import { Option } from 'pages/AdageIframe/app/types'
 import { computeVenueDisplayName } from 'repository/venuesService'
 
 import { DEFAULT_DETAILS_FORM_VALUES } from './constants'
+import { deSerializeDurationMinutes } from './serializers'
 import {
   DetailsFormValues,
   SetDefaultInitialValuesFromOfferProps,
@@ -195,12 +195,6 @@ export function getInitialValuesFromOffer({
   }
 }
 
-export function deSerializeDurationMinutes(durationMinute: number): string {
-  const hours = Math.floor(durationMinute / 60)
-  const minutes = (durationMinute % 60).toString().padStart(2, '0')
-  return `${hours}:${minutes}`
-}
-
 export function setFormReadOnlyFields(
   offer: GetIndividualOfferResponseModel | null,
   isProductBased?: boolean
@@ -228,66 +222,5 @@ export function setFormReadOnlyFields(
     return allFields
   } else {
     return ['categoryId', 'subcategoryId', 'venueId']
-  }
-}
-
-export const serializeExtraData = (formValues: DetailsFormValues) => {
-  return trimStringsInObject({
-    author: formValues.author,
-    gtl_id: formValues.gtl_id,
-    performer: formValues.performer,
-    showType: formValues.showType,
-    showSubType: formValues.showSubType,
-    speaker: formValues.speaker,
-    stageDirector: formValues.stageDirector,
-    visa: formValues.visa,
-    ean: formValues.ean,
-  })
-}
-
-type PostPayload = {
-  description?: string | null
-  durationMinutes?: number
-  extraData?: Record<string, unknown>
-  name: string
-  subcategoryId: string
-  venueId: number
-  productId?: number
-}
-
-export function serializeDetailsPostData(
-  formValues: DetailsFormValues
-): PostPayload {
-  return trimStringsInObject({
-    name: formValues.name,
-    subcategoryId: formValues.subcategoryId,
-    venueId: Number(formValues.venueId),
-    description: formValues.description,
-    durationMinutes: serializeDurationMinutes(formValues.durationMinutes ?? ''),
-    extraData: serializeExtraData(formValues),
-    productId: formValues.productId ? Number(formValues.productId) : undefined,
-    url: formValues.url,
-  })
-}
-
-type PatchPayload = {
-  description?: string | null
-  durationMinutes?: number
-  extraData?: Record<string, unknown>
-  name: string
-  subcategoryId: string
-  url?: string | null
-}
-
-export function serializeDetailsPatchData(
-  formValues: DetailsFormValues
-): PatchPayload {
-  return {
-    name: formValues.name,
-    subcategoryId: formValues.subcategoryId,
-    description: formValues.description,
-    durationMinutes: serializeDurationMinutes(formValues.durationMinutes ?? ''),
-    extraData: serializeExtraData(formValues),
-    url: formValues.url,
   }
 }
